Finally moved cpu into the appropriate file
